+++ includes/class-woocommerce.php
@@ -338,6 +338,11 @@ final class WooCommerce {
 		include_once WC_ABSPATH . 'includes/interfaces/class-wc-queue-interface.php';
 
 		/**
+		 * Core traits.
+		 */
+		include_once WC_ABSPATH . 'includes/traits/trait-wc-item-totals.php';
+
+		/**
 		 * Abstract classes.
 		 */
 		include_once WC_ABSPATH . 'includes/abstracts/abstract-wc-data.php';
